web_scraping/requirements.txt
pymongo
python-dotenv
selenium